dbm_assistant/requirements.txt
pandas==2.0.3
python-dotenv==1.0.1
minio==7.2.12
pyyaml==6.0.2
psycopg2-binary==2.9.10
loguru==0.7.3
minio==7.2.12
fastapi==0.115.6
uvicorn==0.34.0
llama-index==0.12.5
llama-cpp-python==0.3.1
llama-index-embeddings-huggingface==0.4.0
llama-index-vector-stores-milvus==0.4.0